snapshot: add module for ser/deser microVM to/from file
This module exposes a safe API to save and restore microvm internal
structures to, respectively from memory that's mapped to a snapshot
file.

The design is primarily focused on speed, thus it simply saves binary
data straight to file-mapped memory. The snapshot file is used to
also hold file-mapped guest memory.

Signed-off-by: Adrian Catangiu <acatan@amazon.com>