vmm: refactored seccomp filters whitelist
Use one define on both aarch64 and x86_64.

For the syscall defines that are not available on musl+aarch64,
created a module which is aarch64 specific. The defines from the
module are exported on aarch64, while for x86_64 we use the libc
ones. This makes the code uniform on both platforms.

Also fixed the seccomp filters for gnu+aarch64 by whitelisting
SYS_pwait.

Fixes: https://github.com/firecracker-microvm/firecracker/issues/1143

Signed-off-by: Andreea Florescu <fandree@amazon.com>