vmm: separate guest memory file and snapshot
The guest memory is saved in the file specified at PUT
/machine-config, and the remainder of the snapshot goes
to the file specified at InstanceStart.

Hacky and not 100% functional version - the original VM
needs to set the memory file path in /machine-config, and
the resumed VMs use a hardcoded file name (snapshot name +
.mem suffix).

The entire microVM state is stored in a single structure
that is subsequently serialized and deserialized in one go.
This further simplifies the `SnapshotImage` struct and
serialization logic.

Signed-off-by: Alexandra Iordache <aghecen@amazon.com>