integ tests: pause/resume to/from snapshot...
...with devices. This needs a bit of tinkering with the jailer's
network namespaces (disabling them) because clones don't work
well in netns yet. The test adds 2 block devices and a net device.
In the original VM, it mounts the scratch block device and creates
a footprint file. The clone resumes the ssh connection (thus
demonstrating the preservation of net device functionality) and
verifies the marker file (demonstrating that the block device
works and is still mounted).

Signed-off-by: Alexandra Iordache <aghecen@amazon.com>