Added `DeviceConfigs` to snapshot image.
This commit augments the snapshot image so that it contains all the
device configurations that the Vmm uses. The snapshot image will
explicitly contain all of these configurations -- even if it would be
possible to programatically recover some of the information from various
parts of the device state, this would make the snapshot code more
difficult to maintain and to extend. This is done since several
functions in the `Vmm` assume that these configuration are extant
throughout execution.

Moreover, this commit adds code to the vmm to actually use this snapshot
functionality.

Signed-off-by: Tamio-Vesa Nakajima <tamiove@amazon.com>