vmm_config: move under rpc_interface
The vmm_config module contained the configuration definitions
used as part of the Vmm RPC interface. Flattened the two modules
in a single 'rpc_interface' module.

The rpc_interface module now contains the API/RPC actions, the
pre-boot and runtime controllers to apply those actions to the Vmm,
and the resources configuration definitions.

A following commit will move resources.rs in this module at which
point the complete API/RPC interface is isolated/contained and
could be enabled/disabled using features for example.

Signed-off-by: Adrian Catangiu <acatan@amazon.com>