devices: net: simplify RX logic
The RX code path logic had too many twists and turns and was hard
to follow and error-prone.

This commit simplifies it a bit and concentrates most of it in a
single function `net.process_rx()`.

Also added comments that hopefully clear up some ambiguity and
should guide the right usage of internal functions.

Signed-off-by: Adrian Catangiu <acatan@amazon.com>