Shutdown VCPU threads so they can thread::join()
This continues the work from the prior commit toward the goal of
releasing all resources with a clean shutdown.  The VCPU threads in
non-test configurations were running with an unbreakable wait, so
they only way to get past them was with libc::exit().

This commit changes that, along with several other entangled issues.
One was that Drop methods were used to try and clean up the Vmm
and the VcpuHandles, but these were not being called.  The reason
is that there was a cyclic dependency: threads held onto the Vmm,
so the Vmm could not exit VCPU threads in its Drop method (as it
wouldn't be dropped so long as the VCPU threads were around).

Another complexity surrounds defining the difference between an
exited thread (from which an exit event can be read out of the
channel) and a finished thread (which tears down the channel).
Trying to send a message to a thread which has finished will
panic, and some CPUs might be in the exit state while others may
not...and reading the "exit state" by taking the message out of
the channel removes it.  It's a bit complicated.

This commit tries to move toward simplification by being explicit
about the Vmm.stop() call being responsible for exiting the VCPUs,
and making the VcpuHandle's Drop method only do the simple task
of joining an exited thread.

The overloaded cyclical 'exit_evt' is replaced in this commit with
dedicated 'vcpu_exit_event' which trigger Vmm.stop(), and
'teardown_evt' which is internally used to bring down the Vmm.

To avoid cycles, all teardown paths take the following route:
   +------------------------+----------------------------+------------------------+
   |        Vmm             |           Action           |           Vcpu         |
   +------------------------+----------------------------+------------------------+
 1 |                        |                            | vcpu.exit(exit_code)   |
 2 |                        |                            | vcpu.exit_evt.write(1) |
 3 |                        | <--- EventFd::exit_evt --- |                        |
 4 | vmm.stop()             |                            |                        |
 5 |                        | --- VcpuEvent::Finish ---> |                        |
 6 |                        |                            | StateMachine::finish() |
 7 |                        |  ------- EventFd ------+   |                        |
   |                        |  <-----teardown_evt----+   |                        |
 8 | vmm.process_exitable() |                            |                        |
 9 | VcpuHandle::join()     |                            |                        |
10 | vmm.process_exitable() returns Some(exit_code) breaking the main event loop  |
   +------------------------+----------------------------+------------------------+

Vcpu initiated teardown starts from `fn Vcpu::exit()` (step 1).
Vmm initiated teardown starts from `pub fn Vmm::stop()` (step 4).
Because the last step in the cycle is handling EventFd in Subscriber
implementation, the Vmm can also propagate an exit code/reason.

Signed-off-by: Adrian Catangiu <acatan@amazon.com>
Co-authored-by: Cerul Alain <ae@metaeducation.com>