uffd: keep fd open in Firecracker as well
In the happy case we can close the fd since the other process has it
open and is using it to serve us pages.
The problem is that if other process crashes/exits, firecracker
guest memory will simply revert to anon-mem behavior which would lead
to silent errors and UB.

If the fault serving process exits while we hold a copy of the FD as
well, the uffd will still be alive but with no one to serve faults
leading to guest freeze (which is an unfortunate, but explicit and
defined behavior).

Use mem::forget() to keep the uffd open in this process as well.

Signed-off-by: Adrian Catangiu <acatan@amazon.com>